plz fix prod

+++ db/migrate/20161103213652_add_radius_to_posts.rb
@@ -1,5 +1,5 @@
 class AddRadiusToPosts < ActiveRecord::Migration[5.0]
   def change
-    add_column :posts, :max_radius, :double
+    add_column :posts, :max_radius, :decimal
   end
 end
